Set FMC pad and stencil sizes to match Samtec guidelines.
Guidelines are available in e.g.

    http://suddendocs.samtec.com/prints/asp-134604-01.pdf

Solder mask is now 0.89mm diameter, which is significantly larger than before.
I believe this will improve yield because each terminals will "wick up" a
significant amount of paste and can create opens if enough is not available.

Pad sizes were 0.635mm, and are now 0.64mm (this is unlikely to matter.)